Updated UI dark mode color scheme and rating logic
The changes include an update to the color scheme of various components for better visibility and consistency. The background colors have been changed in card, ratingButtons, resultsRow, and resultsTable components. Additionally, dark mode specific border colors have been added to improve visibility in dark mode.

In terms of functionality, the logic for assigning 'orange' class in resultsRow component has been updated to include another condition. Now it will be assigned not only when currentRating equals 'nobut', but also when it equals 'maybe'. This change enhances the flexibility of our rating system.